src/requirements.txt
colour_demosaicing==0.1.6
faiss==1.7.1
numpy==1.21.4
Pillow==8.4.0
scikit_learn==1.0.1
scipy==1.7.2
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tqdm==4.62.3
